fix: address CodeRabbit review — route extraction, Decimal price, rate limiter, reactivity, Intl formatters
- Extract join/fallback/weight logic from route handler into build_heatmap_tickers()
- Change HeatmapTicker.price from float to Decimal (project financial precision rules)
- Wrap yf.download batch call with rate limiter (async with self._limiter)
- Replace hardcoded TTL_HEATMAP with centralized self._cache.ttl_for("heatmap")
- Add reactive timeSince via ticking now ref (30s interval) + computed lastUpdatedText
- Replace toFixed/custom K/M with Intl.NumberFormat (priceFormatter, volumeFormatter, percentFormatter)
- Add keyboard accessibility: role, tabindex, aria-label, Enter/Space handlers, focus-visible styles
- Fix tooltip clamping: Math.max(TOOLTIP_MARGIN, ...) for both edges

+++ src/options_arena/api/schemas.py
@@ -8,6 +8,7 @@ from __future__ import annotations
 
 import math
 from datetime import datetime
+from decimal import Decimal
 from typing import Self
 
 from pydantic import BaseModel, ConfigDict, Field, field_validator, model_validator
@@ -488,7 +489,7 @@ class HeatmapTicker(BaseModel):
     industry_group: str
     market_cap_weight: float
     change_pct: float | None
-    price: float
+    price: Decimal
     volume: int
 
     @field_validator("market_cap_weight")
@@ -502,8 +503,8 @@ class HeatmapTicker(BaseModel):
 
     @field_validator("price")
     @classmethod
-    def _validate_price(cls, v: float) -> float:
-        if not math.isfinite(v):
+    def _validate_price(cls, v: Decimal) -> Decimal:
+        if not math.isfinite(float(v)):
             raise ValueError("price must be finite")
         if v <= 0:
             raise ValueError("price must be positive")
